backend 폴더 밖에 있던 api를 backend 폴더의 app.js 파일로 옮겼습니다. 서버 실행 후 확인까지 마쳤습니다.
